nextjs-lai/pdf-ocr/requirements.txt
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.2.0
openai>=1.0.0
python-dotenv==1.0.1
tqdm==4.66.1
langdetect==1.0.9
pyspellchecker==0.8.1
